Split DSTypeCategory into its own source file
